+++ include/zusi/command.hpp
@@ -15,14 +15,14 @@
 namespace zusi {
 
 enum class Command : uint8_t {
-  None = 0u,     ///< No instruction
-  CvRead = 1u,   ///< Read CV (ack->busy->data)
-  CvWrite = 2u,  ///< Write CV (ack->busy)
-  ZppErase = 4u, ///< ZPP erase (ack->busy)
-  ZppWrite = 5u, ///< ZPP write (ack->busy)
-  Features = 6u, ///< Feature request (ack->busy->data)
-  Exit = 7u,     ///< Exit (ack)
-  Encrypt = 13u  ///< Ask if encryption is supported
+  None = 0x00u,
+  CvRead = 0x01u,
+  CvWrite = 0x02u,
+  ZppErase = 0x04u,
+  ZppWrite = 0x05u,
+  Features = 0x06u,
+  Exit = 0x07u,
+  ZppLcDcQuery = 0x0D
 };
 
 } // namespace zusi