+++ tests/rx/zpp_lc_dc_query.cpp
@@ -2,9 +2,9 @@
 
 using namespace std::chrono_literals;
 
-TEST_F(RxTest, encrypt) {
+TEST_F(RxTest, zpp_lc_dc_query) {
   zusi::Packet packet{
-    std::to_underlying(zusi::Command::Encrypt), 1u, 2u, 3u, 4u};
+    std::to_underlying(zusi::Command::ZppLcDcQuery), 1u, 2u, 3u, 4u};
 
   EXPECT_CALL(_mock, receiveByte())
     .WillOnce(Return(packet[0uz]))
@@ -15,7 +15,7 @@ TEST_F(RxTest, encrypt) {
     .WillOnce(Return(zusi::crc8(packet)))
     .WillOnce(Return(zusi::resync_byte))
     .WillRepeatedly(Return(std::nullopt));
-  EXPECT_CALL(_mock, gpio()).Times(1);
+  EXPECT_CALL(_mock, gpioOutput()).Times(1);
   EXPECT_CALL(_mock, waitClock(_)).WillRepeatedly(Return(true));
   EXPECT_CALL(_mock, loadCodeValid(ElementsAre(1u, 2u, 3u, 4u)))
     .WillOnce(Return(true));