+++ tests/tx/zpp_update.cpp
@@ -14,10 +14,10 @@ using ::Mbps::_1_364;
 using ::Mbps::_1_807;
 using ::zusi::resync_byte;
 
-inline constexpr uint32_t mock_addr{0x000000FFu};
+inline constexpr uint32_t _addr{0x000000FFu};
 inline constexpr uint8_t mock_val{0x0Fu};
 
-TEST_F(TxTest, write_zpp_no_ack_valid) {
+TEST_F(TxTest, zpp_update_no_ack_valid) {
   std::span<uint8_t const> mock_vals{&mock_val, 1u};
 
   InSequence seq;
@@ -27,11 +27,11 @@ TEST_F(TxTest, write_zpp_no_ack_valid) {
   EXPECT_CALL(_mock, readData()).WillOnce(Return(true)); // not ACK valid
   EXPECT_CALL(_mock, spiMaster());
 
-  ASSERT_FALSE(_mock.writeZpp(mock_addr, mock_vals))
+  ASSERT_FALSE(_mock.writeZpp(_addr, mock_vals))
     << "Should abort if not ACK valid";
 }
 
-TEST_F(TxTest, write_zpp_nak) {
+TEST_F(TxTest, zpp_update_nak) {
   std::span<uint8_t const> mock_vals{&mock_val, 1u};
 
   InSequence seq;
@@ -42,11 +42,10 @@ TEST_F(TxTest, write_zpp_nak) {
   EXPECT_CALL(_mock, readData()); // NAK
   EXPECT_CALL(_mock, spiMaster());
 
-  ASSERT_FALSE(_mock.writeZpp(mock_addr, mock_vals))
-    << "Should abort after NAK";
+  ASSERT_FALSE(_mock.writeZpp(_addr, mock_vals)) << "Should abort after NAK";
 }
 
-TEST_F(TxTest, write_zpp_busy_wait) {
+TEST_F(TxTest, zpp_update_busy_wait) {
   std::span<uint8_t const> mock_vals{&mock_val, 1u};
 
   InSequence seq;
@@ -59,11 +58,11 @@ TEST_F(TxTest, write_zpp_busy_wait) {
   EXPECT_CALL(_mock, readData()).WillOnce(Return(true)); // Busy End
   EXPECT_CALL(_mock, spiMaster());
 
-  ASSERT_TRUE(_mock.writeZpp(mock_addr, mock_vals))
+  ASSERT_TRUE(_mock.writeZpp(_addr, mock_vals))
     << "Should continue after no longer busy";
 }
 
-TEST_F(TxTest, write_zpp_ack) {
+TEST_F(TxTest, zpp_update_ack) {
   std::span<uint8_t const> mock_vals{&mock_val, 1u};
 
   InSequence seq;
@@ -79,6 +78,6 @@ TEST_F(TxTest, write_zpp_ack) {
   EXPECT_CALL(_mock, readData()).WillOnce(Return(true)); // Busy
   EXPECT_CALL(_mock, spiMaster());
 
-  auto tmp{_mock.writeZpp(mock_addr, mock_vals)};
+  auto tmp{_mock.writeZpp(_addr, mock_vals)};
   ASSERT_TRUE(tmp) << "Should return true if command is correct";
 }